fix: moving config.json to config/
