+++ .gitignore
@@ -1,5 +1,5 @@
 .env
-config.json
+config/config.json
 config/database.json
 config/accounts.json
 config/key.json
