+++ entities/config/DatabaseConfig.ts
@@ -1,4 +1,4 @@
-import json from "@/config/config.json" with { type: "json" };
+import json from "@/config/database.json" with { type: "json" };
 
 interface DatabaseConfigType {
 	username: string;
